v1.1 release merge from develop

+++ Source/Presets/PresetComp.h
@@ -9,24 +9,31 @@
 */
 
 #pragma once
-#include "../UI/LookAndFeel.h"
 
-struct PresetComp : Component
+struct PresetComp : Component, private Timer
 {
     PresetComp(AudioProcessorValueTreeState& vts) : manager(vts)
     {
         addAndMakeVisible(box);
         box.setJustificationType(Justification::centredLeft);
+        box.setTextWhenNothingSelected("Presets");
 
         loadPresets();
 
         addChildComponent(editor);
         editor.setJustification(Justification::centredLeft);
+
+        startTimerHz(2);
+    }
+
+    ~PresetComp()
+    {
+        stopTimer();
     }
 
     void loadPresets()
     {
-        PopupMenu::Item save{ "save" }, saveAs{ "save as" }, copy{ "copy state" }, paste{"paste state"};
+        PopupMenu::Item save{ "save" }, saveAs{ "save as" }, copy{ "copy state" }, paste{"paste state"}, presetDir{"open preset folder"};
 
         auto menu = box.getRootMenu();
         menu->clear();
@@ -34,10 +41,12 @@ struct PresetComp : Component
         save.setTicked(false);
         save.action = [&] { savePreset(); };
         menu->addItem(save);
+
         saveAs.setID(1002);
         saveAs.setTicked(false);
         saveAs.action = [&] { savePresetAs(); };
         menu->addItem(saveAs);
+
         copy.setID(1003);
         copy.setTicked(false);
         copy.action = [&]
@@ -45,6 +54,7 @@ struct PresetComp : Component
             SystemClipboard::copyTextToClipboard(manager.getStateAsString());
         };
         menu->addItem(copy);
+
         paste.setID(1004);
         paste.setTicked(false);
         paste.action = [&]
@@ -52,6 +62,14 @@ struct PresetComp : Component
             manager.setStateFromString(SystemClipboard::getTextFromClipboard());
         };
         menu->addItem(paste);
+
+        presetDir.setID(1005);
+        presetDir.setTicked(false);
+        presetDir.action = [&]
+        {
+            manager.userDir.getParentDirectory().startAsProcess();
+        };
+        menu->addItem(presetDir);
         menu->addSeparator();
 
         auto presets = manager.loadFactoryPresetList();
@@ -81,9 +99,18 @@ struct PresetComp : Component
         box.setText(currentPreset, NotificationType::dontSendNotification);
     }
 
+    void setPresetWithChange(const String& newPreset) noexcept
+    {
+        currentPreset = newPreset;
+        box.setText(currentPreset, NotificationType::sendNotificationAsync);
+    }
+
     void savePreset() noexcept
     {
+        if (box.getText() == "" || currentPreset == "")
+            return;
         manager.savePreset(currentPreset, manager.userDir);
+        setCurrentPreset(currentPreset);
     }
 
     void savePresetAs() noexcept
@@ -91,68 +118,69 @@ struct PresetComp : Component
         editor.setVisible(true);
         editor.toFront(true);
         editor.setText("Preset Name", false);
+        editor.grabKeyboardFocus();
         editor.setHighlightedRegion({ 0, 12 });
 
         editor.onFocusLost = [&]
         {
-#if JUCE_WINDOWS
-            editor.clear(); editor.setVisible(false);
-#elif JUCE_MAC
-            editor.onReturnKey();
-#endif
+            editor.clear();
+            editor.setVisible(false);
         };
+
         editor.onEscapeKey = [&] { editor.clear(); editor.setVisible(false); };
 
         editor.onReturnKey = [&]
         {
             auto name = editor.getText();
+            if (name == "")
+            {
+                box.setText("Enter a name!");
+                return;
+            }
             editor.clear();
             editor.setVisible(false);
 
             if (manager.savePreset(name, manager.userDir)) {
                 loadPresets();
-                box.setText(name, NotificationType::sendNotificationSync);
-                currentPreset = name;
+                setPresetWithChange(name);
             }
             else
                 box.setText("invalid name!", NotificationType::dontSendNotification);
         };
-        
     }
 
-    void valueChanged(bool presetModified) noexcept
+    void valueChanged() noexcept
     {
         auto id = box.getSelectedId();
         auto idx = box.getSelectedItemIndex();
         auto preset = box.getItemText(idx);
 
-        if (!presetModified) {
-            if (id < 1000 && id > 0) {
+        if (id < 1000 && id > 0) {
 
-                if (id <= factoryPresetSize) {
-                    if (manager.loadPreset(preset, true))
-                        box.setText(preset, NotificationType::sendNotificationSync);
-                    else
-                        box.setText("preset not found", NotificationType::dontSendNotification);
-                }
-                else {
-                    if (manager.loadPreset(preset, false))
-                        box.setText(preset, NotificationType::sendNotificationSync);
-                    else
-                        box.setText("preset not found", NotificationType::dontSendNotification);
-                }
-
-                currentPreset = preset;
+            if (id <= factoryPresetSize) {
+                if (manager.loadPreset(preset, true))
+                    box.setText(preset, NotificationType::sendNotificationSync);
+                else
+                    box.setText("preset not found", NotificationType::dontSendNotification);
             }
-            else
-                box.setText(currentPreset, NotificationType::dontSendNotification);
+            else {
+                if (manager.loadPreset(preset, false))
+                    box.setText(preset, NotificationType::sendNotificationSync);
+                else
+                    box.setText("preset not found", NotificationType::dontSendNotification);
+            }
+
+            currentPreset = preset;
         }
         else
-            box.setText(preset, NotificationType::dontSendNotification);
+            box.setText(currentPreset, NotificationType::dontSendNotification);
     }
 
-    void paint(Graphics& g) override
-    {}
+    void timerCallback() override
+    {
+        if (manager.hasStateChanged() && currentPreset != "")
+            box.setText(currentPreset + "*", NotificationType::dontSendNotification);
+    }
 
     void resized() override
     {
